Properly package
Implement this code as a proper Python package that can be installed
with setuptools/pip.

Use the click package to implement a `bing_geocoder` command that
will get installed as an executable instead of using `main()`
methods within the various modules.

Catch import errors so boto and sendgrid are not neccessary if all
you want to do is geocode from another application.